fix(class): Fix class issue
