+++ app/router.js
@@ -61,7 +61,7 @@ Router.map(function() {
       });
     }
   });
-
+  this.route('class', {path: '/classes/:class'})
 });
 
 export default Router;
