+++ app/routes/project-version/namespaces/namespace/index.js
@@ -1,5 +1,5 @@
 import Ember from 'ember';
-import HashRedirectMixin from '../../../../mixins/hash-redirect';
+import HashRedirectMixin from 'ember-api-docs/mixins/hash-redirect';
 
 export default Ember.Route.extend(HashRedirectMixin, {
   templateName: 'class-index'
